session: de-genericise internal query/exec functions
The query/execute/batch statements are generic over the statement.
They started by converting the statement to corresponding type
(query/execute/batch) and then continued without the need for generics.
However, those functions used to be non-trivial and would have to be
monomorphised for every type of the arguments passed to the method,
increasing compilation time more than necessary.

Now that most of the implementation was moved to do_query etc. methods,
we can restrict the generic part to the public query/execute/batch
methods which convert the input statement to required type and then call
the non-generic do_query etc. methods. This commit does just that -
de-genericises do_query and friends, while leaving query and friends
generic as they used to.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>