transport: introduce new QueryResult
(Re)-introduces QueryResult. It is quite similar to the old
(Legacy)QueryResult, but it keeps rows and metadata  in an unparsed
state (via RawRows) and has methods that allow parsing the contents
using the new API.

Helper method names are similar to what the old QueryResult had, just
with the `_typed` suffix dropped - as now it is always required to
provide the type of rows when parsing them, this suffix sounded
redundant.

There is one crucial change to the API. Motivation is as follows:
1) `QueryResult` can represent a non-Rows response, so every
   rows-related operation on `QueryResult` may return "NonRowsResponse"
   error, which is inconvenient;
2) `QueryResult` is an owned type, so it cannot deserialize metadata
   in the borrowed flavour (i.e., using strings from the frame bytes
   directly) and it must allocate metadata (mainly ColumnSpecs) on the
   heap.
The solution for both is to extract a new struct, `RowsDeserializer`,
which is parametrized by a lifetime and hence can borrow metadata from
the frame. Moreover, one has to handle "NonRowsResponse" error only
once, upon `RowsDeserializer` creation. All further methods (`rows(),
`single_row()`, etc.) may no longer fail with that error, which provides
a clean step of conversion from any Result frame to Result:Rows frame.

The drawback is that now there is a new call required in the call chain
to deserialize a result, namely `.row_deserializer()`.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>