cluster: make Cluster cheaply cloneable
The Cluster struct by itself only serves as a facade for the
ClusterWorker, i.e. it has channels that allow sending requests to the
worker, receives the ClusterData via the `data` field etc. Apart from
the `_worker_handle` field, all other fields are cloneable. Two tasks
working on two copies of the same Cluster object should behave the same
as if they shared and operated on a single Cluster object (e.g. via
Arc<Cluster>).

This commit makes the Cluster object cloneable - the `_worker_handle` is
shared via an Arc. This will be very useful in the next commit - we will
do a similar thing for the GenericSession object.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>