iterator: adjust to the new deserialization framework
This commit finishes the work related to adjusting the iterators module
to the new deserialization framework.

The equivalent of the old RowIterator is now RawIterator (notice the
vowel change). Despite the name, it cannot actually be iterated on, as
it does not have any information about the column types. After calling
the `into_typed()` method it gets converted into TypedRowIterator that
can actually be iterated on.

Unfortunately, due to the limitations of the Stream trait, currently the
TypedRowIterator cannot deserialize types that need to borrow from the
serialized frame contents.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>