iterator: adjust to the new deserialization framework
This commit makes the RowIteratorWorker pass raw rows to the main tokio
task, instead of the eagerly deserialized Rows.

The equivalent of the old RowIterator is now RawIterator (notice a
letter change). Despite the name, it cannot actually be conveniently
iterated on, as it does not have any information about the column types.
It exposes a `next()` method for deserializing consecutive
`ColumnIterator`s. Users can manually perform deserialization using this
method directly, but the preferred (typed) API will be added in the next
commit.

The legacy iterators are preserved by wrapping around RawIterator.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>