session,iterator: no longer record rows size & count
After the deserialization refactor with lazy result metadata
deserialization, we will no longer have access to rows serialized size
and rows count in Session and in RowIteratorWorker. Therefore, traces
can no longer record that information.

A further commit in this PR brings back serialized size to the span,
but there is an important change: now the size is of both raw metadata
and raw rows; before, metadata was not accounted.