session: prepare on every node, not shard
Scylla prepares the statement on every shard when processing PREPARE
request:
https://github.com/scylladb/scylladb/blob/8f0d0daf53397aa68312571ab9d01d8b75cd1770/transport/server.cc#L1104-L1114.

At the same time, driver unnecessarily tries to prepare the statement on
all connections, which means connections to every shard, so possibly
multiple connections to every node.

This commit makes the driver prepare the statement only on a single
(random) connection to every node.

There is one catch: with the new logic, we might sometimes fail
preparation even though we could succeed if tried on different shards
(some shards might be, for example, overloaded). This issue is solved
in further commits.