request: introduce BatchV2
BatchV2 is dedicated to support bound batches, i.e., batches that are
made up of bound statements only. Its main difference from the existing
Batch structure is that it stores the serialized statements and values
directly as a byte array, which is in line with how bound statements are
handled elsewhere in the driver. This is in contrast to the existing
Batch structure, which stores statements and values in a non-serialized
form and requires additional serialization steps when sending the batch
to the server.

BatchV2 therefore involves more overhead, as it requires serializing
the statements and values into a byte array before creating the
BatchV2 instance, which will then be once again copied into the final
request buffer when the request is serialized. However, this design
greatly simplifies the handling of bound batches, as it eliminates the
need for complex serialization logic within the Batch structure itself.

Co-Authored-By: Andres Medina <andresnrx@gmail.com>